add template for C++

+++ tester/tester.sh
@@ -192,6 +192,8 @@ fi
 
 COMPILE_BEGIN_TIME=$(($(date +%s%N)/1000000));
 
+
+
 ########################################################################################################
 ############################################ COMPILING JAVA ############################################
 ########################################################################################################
@@ -293,7 +295,27 @@ if [ "$EXT" = "c" ] || [ "$EXT" = "cpp" ]; then
 		COMPILER="g++ -std=c++98"
 	fi
 	EXEFILE="s_$(echo $FILENAME | sed 's/[^a-zA-Z0-9]//g')" # Name of executable file
-	cp $PROBLEMPATH/$UN/$FILENAME.$EXT code.c
+
+	if [ -f "$PROBLEMPATH/template.cpp" ]; then
+		t="$PROBLEMPATH/template.cpp"
+		f=$PROBLEMPATH/$UN/$FILENAME.$EXT
+		banned=`sed -n -e '/###Begin banned keyword/,/###End banned keyword/p' $t | sed -e '1d' -e '$d'`
+		code=`sed -e '1,/###End banned keyword/d' $t`
+		#echo "$banned"
+		#echo "$code"
+		while read -r line
+		do
+			#echo grep -q "$line" $f
+			if grep -q "$line" $f ;then
+				echo "$line is banned" >> cerr
+				NEEDCOMPILE=0
+			fi
+		done <<< "$banned"
+		echo "$code" | sed -e "/###INSERT CODE HERE/r $f" -e '/###INSERT CODE HERE/d' > code.c
+	else
+		cp $PROBLEMPATH/$UN/$FILENAME.$EXT code.c
+	fi
+
 	shj_log "Compiling as $EXT"
 	if $SANDBOX_ON; then
 		shj_log "Enabling EasySandbox"
@@ -304,25 +326,31 @@ if [ "$EXT" = "c" ] || [ "$EXT" = "cpp" ]; then
 			SANDBOX_ON=false
 		fi
 	fi
-	if $C_SHIELD_ON; then
-		shj_log "Enabling Shield For C/C++"
-		# if code contains any 'undef', raise compile error:
-		if tr -d ' \t\n\r\f' < code.c | grep -q '#undef'; then
-			echo 'code.c:#undef is not allowed' >cerr
-			EXITCODE=110
+
+	if [ $NEEDCOMPILE -eq 0 ]; then
+		EXITCODE=1
+	else
+		if $C_SHIELD_ON; then
+			shj_log "Enabling Shield For C/C++"
+			# if code contains any 'undef', raise compile error:
+			if tr -d ' \t\n\r\f' < code.c | grep -q '#undef'; then
+				echo 'code.c:#undef is not allowed' >cerr
+				EXITCODE=110
+			else
+				cp ../shield/shield.$EXT shield.$EXT
+				cp ../shield/def$EXT.h def.h
+				# adding define to beginning of code:
+				echo '#define main themainmainfunction' | cat - code.c > thetemp && mv thetemp code.c
+				$COMPILER shield.$EXT $C_OPTIONS $C_WARNING_OPTION -o $EXEFILE >/dev/null 2>cerr
+				EXITCODE=$?
+			fi
 		else
-			cp ../shield/shield.$EXT shield.$EXT
-			cp ../shield/def$EXT.h def.h
-			# adding define to beginning of code:
-			echo '#define main themainmainfunction' | cat - code.c > thetemp && mv thetemp code.c
-			$COMPILER shield.$EXT $C_OPTIONS $C_WARNING_OPTION -o $EXEFILE >/dev/null 2>cerr
+			mv code.c code.$EXT
+			$COMPILER code.$EXT $C_OPTIONS $C_WARNING_OPTION -o $EXEFILE >/dev/null 2>cerr
 			EXITCODE=$?
 		fi
-	else
-		mv code.c code.$EXT
-		$COMPILER code.$EXT $C_OPTIONS $C_WARNING_OPTION -o $EXEFILE >/dev/null 2>cerr
-		EXITCODE=$?
 	fi
+
 	COMPILE_END_TIME=$(($(date +%s%N)/1000000));
 	shj_log "Compiled. Exit Code=$EXITCODE  Execution Time: $((COMPILE_END_TIME-COMPILE_BEGIN_TIME)) ms"
 	if [ $EXITCODE -ne 0 ]; then
